+++ modules/common/src/errors.ts
@@ -84,3 +84,9 @@ export class ProposalCreationError extends Error {
     super("Failed to create proposal");
   }
 }
+
+export class MissingExecPermissionError extends Error {
+  constructor() {
+    super("No plugin requests EXECUTE_PERMISSION");
+  }
+}